+++ CHANGELOG.md
@@ -8,7 +8,10 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
 Once we reach the v1.0 release, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing so far
+- Allow adapters to implement the optional `ReactionAwareAdapter` interface if they support emoji reactions
+- Add new `reactions` package which contains a compiled list of all officially supported reactions
+- Components may now return the new `ErrNotImplemented` if they do not support a feature
+- Add new `reactions.Event` that may be emitted by an Adapter so users can listen for it 
 
 ## [v0.8.0] - 2019-04-21
 - Make Auth.Grant(…) idempotent and do not unnecessarily add smaller scopes