+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
 Once we reach the v1.0 release, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Nothing so far
+- Add new `FinishEventContent(…)` function to finish event processing with multiple handlers early
+- Message handlers registered via `Bot.Respond(…)` now abort early if the pattern matches
+- This allows users to specify a default response when nothing else matches (see #25)
 
 ## [v0.9.0] - 2019-10-22
 - Add `Auth.Users()` and `Auth.UserPermissions(…)` functions to allow retrieving all users as well as users permissions.